requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
cryptography==2.3.1
DateTime==4.2
Django==2.1.1
freeze==1.0.10
idna==2.7
lxml==4.2.4
mysql-connector==2.1.6
netaddr==0.7.19
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.19.1
selenium==3.14.0
six==1.11.0
urllib3==1.23
zope.interface==4.5.0
